ECS: Added the basics of events Remade tests folder PATCH
