+++ includes/Core/Core.hpp
@@ -4,6 +4,9 @@
 #include "App.hpp"
 #include "Clock.hpp"
 #include "Components/Component.hpp"
+#include "Events/Event.hpp"
+#include "Events/EventHandler.hpp"
+#include "Events/EventsManager.hpp"
 #include "Systems/GenericSystem.hpp"
 #include "Systems/System.hpp"
 #endif /* !CORE_HPP_ */
