+++ test/TestsWorld.cpp
@@ -7,49 +7,6 @@
 #include "ECS.hpp"
 #include <catch2/catch_test_macros.hpp>
 
-TEST_CASE("App", "[App]")
-{
-    Engine::App app;
-    auto world = std::make_unique<Engine::Core::World>();
-
-    SECTION("Add a world")
-    {
-        app.addWorld(0, std::move(world));
-        REQUIRE(app[0] != nullptr);
-    }
-    SECTION("Add a world with a key already used")
-    {
-        app.addWorld(0, std::move(world));
-        REQUIRE_THROWS_AS(app.addWorld(0, std::move(world)), Engine::AppExceptionKeyAlreadyExists);
-    }
-    SECTION("Get a world with a key not used")
-    {
-        REQUIRE_THROWS_AS(app[0], Engine::AppExceptionKeyNotFound);
-    }
-    SECTION("Get a world with a key used")
-    {
-        app.addWorld(0, std::move(world));
-        REQUIRE_NOTHROW(app[0]);
-    }
-    SECTION("Get a world with a key used and check if it's the same")
-    {
-        app.addWorld(0, std::move(world));
-        REQUIRE(app[0] == app[0]);
-    }
-    SECTION("Get a world with a key used and check if it's not the same")
-    {
-        app.addWorld(0, std::move(world));
-        app.addWorld(1, std::move(world));
-        REQUIRE(app[0] != app[1]);
-    }
-    SECTION("Get a world with a key used and check if it's not the same after a move")
-    {
-        app.addWorld(0, std::move(world));
-        auto tmp = std::move(app[0]);
-        REQUIRE(app[0] != tmp);
-    }
-}
-
 struct hp1 : public Engine::Component
 {
     public:
@@ -193,118 +150,5 @@ TEST_CASE("World", "[World]")
         REQUIRE(hp1Comp2.hp == hps - 4);
         REQUIRE(hp2Comp2.maxHp == hps - 8);
         REQUIRE(hp2Comp3.maxHp == hps);
-    }
-}
-
-TEST_CASE("Query")
-{
-    Engine::Core::World world;
-    constexpr int hps = 10;
-
-    struct Player : public Engine::Component
-    {};
-
-    SECTION("getAllEntities")
-    {
-        world.registerComponents<hp1, hp2, Player>();
-
-        world.createEntity();
-        world.createEntity();
-
-        world.addComponentToEntity(0, hp1 {hps});
-        world.addComponentToEntity(1, hp1 {hps});
-        world.addComponentToEntity(0, hp2 {hps});
-        world.addComponentToEntity(1, hp2 {hps});
-        world.addComponentToEntity(0, Player {});
-
-        auto query = world.query<hp1, hp2, Player>().getAllEntities();
-        REQUIRE(query.size() == 1);
-        REQUIRE(query[0] == 0);
-    }
-
-    SECTION("getAllEntities 2")
-    {
-        world.registerComponents<hp1, hp2, Player>();
-
-        world.createEntity();
-        world.createEntity();
-
-        world.addComponentToEntity(0, hp1 {hps});
-        world.addComponentToEntity(1, hp1 {hps});
-        world.addComponentToEntity(0, hp2 {hps});
-        world.addComponentToEntity(1, hp2 {hps});
-        world.addComponentToEntity(0, Player {});
-        world.addComponentToEntity(1, Player {});
-
-        auto query = world.query<hp1, hp2, Player>().getAllEntities();
-        REQUIRE(query.size() == 2);
-        REQUIRE(query[0] == 0);
-        REQUIRE(query[1] == 1);
-    }
-
-    SECTION("getAllEntities 3")
-    {
-        world.registerComponents<hp1, hp2, Player>();
-
-        world.createEntity();
-        world.createEntity();
-        world.createEntity();
-
-        world.addComponentToEntity(0, hp1 {hps});
-        world.addComponentToEntity(1, hp1 {hps});
-        world.addComponentToEntity(2, hp1 {hps});
-        world.addComponentToEntity(0, hp2 {hps});
-        world.addComponentToEntity(1, hp2 {hps});
-        world.addComponentToEntity(2, hp2 {hps});
-        world.addComponentToEntity(0, Player {});
-        world.addComponentToEntity(2, Player {});
-
-        auto query = world.query<hp1, hp2, Player>().getAllEntities();
-        REQUIRE(query.size() == 2);
-        REQUIRE(query[0] == 0);
-        REQUIRE(query[1] == 2);
-    }
-
-    SECTION("getComponentsOfEntity")
-    {
-        world.registerComponents<hp1, hp2, Player>();
-
-        world.createEntity();
-
-        world.addComponentToEntity(0, hp1 {hps});
-        world.addComponentToEntity(0, hp2 {hps});
-        world.addComponentToEntity(0, Player {});
-
-        auto query = world.query<hp1, hp2, Player>().getComponentsOfEntity(0);
-        const hp1 &hp1Comp = std::get<0>(query);
-        const hp2 &hp2Comp = std::get<1>(query);
-        const Player &playerComp = std::get<2>(query);
-
-        REQUIRE(hp1Comp.hp == hps);
-        REQUIRE(hp2Comp.maxHp == hps);
-        REQUIRE_NOTHROW(playerComp);
-    }
-
-    SECTION("getAll")
-    {
-        world.registerComponents<hp1, hp2, Player>();
-
-        auto idx1 = world.createEntity();
-        world.createEntity();
-        auto idx2 = world.createEntity();
-
-        world.addComponentToEntity(0, hp1 {hps});
-        world.addComponentToEntity(1, hp1 {hps});
-        world.addComponentToEntity(2, hp1 {hps});
-        world.addComponentToEntity(0, hp2 {hps});
-        world.addComponentToEntity(1, hp2 {hps});
-        world.addComponentToEntity(2, hp2 {hps});
-        world.addComponentToEntity(0, Player {});
-        world.addComponentToEntity(2, Player {});
-
-        auto query = world.query<hp1, hp2, Player>().getAll();
-        REQUIRE(query.size() == 2);
-        REQUIRE(std::get<0>(query[0]) == idx1);
-        REQUIRE(std::get<0>(query[1]) == idx2);
     }
 }
